Support --filter for Prometheus recording
Move MetricFilter to the shared recorder module so both Prometheus and
Valkey converters can use it. The same JSON array format works for both
sources — entries match against the native metric name of each source
(Prometheus metric names or Redis INFO field names).
